django-backend/ukshub/requirements.txt
asgiref==3.4.1
attrs==21.2.0
charset-normalizer==2.0.7
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.20.2
idna==3.3
inflection==0.5.1
jsonschema==4.2.1
Markdown==3.3.4
psycopg2==2.9.1
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
